requirements.txt
black
fastapi
functions-framework
httpx
mangum
slowapi
python-dotenv
pytest
pytest-asyncio
pytest-mock